food-similarity-scoring/pyproject.toml
[project]
name = "food-similarity"
version = "0.1.0"
description = "Food similarity search tool for diet recommendation research"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.51.0",
    "sentence-transformers>=2.7.0",
    "faiss-cpu>=1.9.0",
    "polars>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "requests>=2.32.5",
    "curl-cffi>=0.13.0",
    "pandas>=3.0.1",
    "tqdm>=4.67.3",
    "beautifulsoup4>=4.14.3",
    "openai>=1.0.0",
]

[project.scripts]
build-index = "scripts.build_index:main"
